Added plan for a fake bank project
This is a self challenge to emulate the fake bank page by Kitboga. The
idea is to eventually be able to use it on stream for my own scambaiting
purposes as well as to practice js/HTML. This should be good for
utilizing objects, having js and HTML interact with each other, editing
html via js and general project creation/maintenance.