Finished CSS autogeneration!
Grid can now be set to any dimensions you want and both html and css
will be generated.
  An ID is generated and added to each button created by the HTML
  The css grid is created based on these IDs and column/row size

Should be ready to start doing things in the DataContainer
Think about whether or not we want each button to create it's own
object. We can then fit the on click functions inside of each one and
they only have to calculat touching buttons once, and can save them from
then on. Then it is only a matter of looking into the saved arrray for
flag/bomb and checking. Look into budgety delete buttons for info if
needed.