Started async functions
Having a really hard time getting a weather api to connect properly. I
may have to do the lesson without it but don't like the idea of not
practicing. crossorigin.me is down cors-anywhere hasn't worked and after
signing up and getting a key for darksky that's not working for me
either.